Finalizado el metodo GET.
Se soluciono el error en los generos. Ahora muestra que generos tiene un libro.

+++ backend/CRUD-books/java/src/main/java/com/example/java/model/Author.java
@@ -12,7 +12,7 @@ public class Author {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_author;
+    private int id_author;
     private String first_name;
     private String last_name;
 }
